feat: remove ugly glue code from rapidjson.
