+++ src/config/AppConfigStore.h
@@ -1,25 +1,10 @@
 #pragma once
 
+#include "AppConfigTypes.h"
+
 #include <filesystem>
 #include <memory>
-
-#include <parakeet-crypto/decryption/DecryptionManager.h>
-
 namespace parakeet_wx::config {
-
-struct GeneralConfig {
-  int thread_count;
-  std::string locale;
-};
-
-struct DesktopConfig {
-  GeneralConfig general;
-};
-
-struct AppConfig {
-  DesktopConfig desktop;
-  parakeet_crypto::decryption::config::DecryptionConfig decryption;
-};
 
 class AppConfigStore {
  public:
