+++ vcpkg.json
@@ -7,6 +7,7 @@
     "boost-filesystem",
     "boost-system",
     "boost-thread",
+    "nlohmann-json",
     "rapidjson",
     "wxwidgets"
   ]
